Move files back out of controller
I tried moving the files into the controller, creating an index file, a router, and a 404 page and but the files don't work. I got the index page to view but for the other pages the url shows the controller.php file rather than loading the view. When the controller still calls the view. At this point in the project i may just forget about a router and work on the interface. Which is a bummer.